Consolidate the HLS test helpers and fixtures
All shared HLS test scaffolding now lives in hls_test_helpers_test.go
instead of being spread across the file that happened to need it first:
the cleanup-blocking helpers, insertTestHLSMovieFixture,
sanitizeTestPathComponent, testIntPtr, the playback-session ID constants
and stubKeyframeFfprobe.

Added:
- safeRemuxFixture / unsafeRemuxFixture / transcodeFixture plus
  hlsRunPlan(fixture), replacing 15 copy-pasted WriteFiles closures.
- newHLSTestHandler, which wires the three personal HLS routes behind the
  session middleware exactly as routes.go does. Six tests hand-built that
  router; they switch over in the next commit.
- noMetadataProbe, so stubKeyframeFfprobe stops hand-writing two
  error-returning methods purely to satisfy ffprobe.FfprobeInterface.
  Mirror of noKeyframeProbe in movies_scanner_test.go.

Coverage holds at 69.7%.
